feat: big-map: drag as system
